Refactor imports and exports per ESM
