+++ packages/lib-user/dev/components/App/index.js
@@ -1,1 +1,1 @@
-export { default } from './App'
+export { default } from './App.js'
