+++ packages/lib-user/package.json
@@ -16,6 +16,7 @@
     }
   },
   "sideEffects": false,
+  "type": "module",
   "repository": "https://github.com/zooniverse/front-end-monorepo.git",
   "bugs": "https://github.com/zooniverse/front-end-monorepo/issues",
   "scripts": {
