+++ packages/lib-user/src/components/GroupStats/index.js
@@ -1,1 +1,1 @@
-export { default } from './GroupStats'
+export { default } from './GroupStats.js'
