+++ packages/lib-user/src/components/index.js
@@ -1,0 +1,3 @@
+export { default as GroupStats } from './GroupStats/index.js'
+
+export { default as UserStats } from './UserStats/index.js'
