+++ packages/lib-user/src/hooks/index.js
@@ -1,4 +1,4 @@
-export { default as useGroupStats } from './useGroupStats'
-export { default as usePanoptesUser } from './usePanoptesUser'
-export { default as usePanoptesUserGroup } from './usePanoptesUserGroup'
-export { default as useUserStats } from './useUserStats'
+export { default as useGroupStats } from './useGroupStats.js'
+export { default as usePanoptesUser } from './usePanoptesUser.js'
+export { default as usePanoptesUserGroup } from './usePanoptesUserGroup.js'
+export { default as useUserStats } from './useUserStats.js'
