+++ packages/lib-user/src/hooks/usePanoptesUser.js
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react'
 import { auth } from '@zooniverse/panoptes-js'
 
-import { getBearerToken } from '@utils'
+import { getBearerToken } from '@utils/index.js'
 
 async function fetchPanoptesUser(authClient) {
   try {
