+++ packages/lib-user/src/hooks/usePanoptesUserGroup.js
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react'
 import { panoptes } from '@zooniverse/panoptes-js'
 
-import { getBearerToken } from '@utils'
+import { getBearerToken } from '@utils/index.js'
 
 // TODO: refactor with SWR
